feat: Implemented Drill Head multiblock

+++ src/main/java/the_fireplace/caterpillar/common/block/entity/DrillHeadBlockEntity.java
@@ -2,25 +2,20 @@ package the_fireplace.caterpillar.common.block.entity;
 
 import net.minecraft.core.BlockPos;
 import net.minecraft.network.chat.Component;
-import net.minecraft.sounds.SoundEvent;
 import net.minecraft.sounds.SoundEvents;
 import net.minecraft.sounds.SoundSource;
-import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.state.BlockState;
-import net.minecraft.world.level.block.state.properties.DoubleBlockHalf;
 import the_fireplace.caterpillar.Caterpillar;
 import the_fireplace.caterpillar.common.block.entity.util.InventoryBlockEntity;
 import the_fireplace.caterpillar.common.container.DrillHeadContainer;
 import the_fireplace.caterpillar.core.init.BlockEntityInit;
 
 import static net.minecraft.world.level.block.HorizontalDirectionalBlock.FACING;
-// import static the_fireplace.caterpillar.common.block.DrillHeadBlock.HALF;
+import static the_fireplace.caterpillar.common.block.DrillHeadBlock.POWERED;
 
 public class DrillHeadBlockEntity extends InventoryBlockEntity {
 
-    // If lever is on, then drill head is powered.
-    private boolean isPowered;
     private int ticks;
 
     public static final Component TITLE = Component.translatable(
@@ -38,12 +33,11 @@ public class DrillHeadBlockEntity extends InventoryBlockEntity {
     public DrillHeadBlockEntity(BlockPos pos, BlockState state) {
         super(BlockEntityInit.DRILL_HEAD.get(), pos, state, DrillHeadContainer.SLOT_SIZE);
 
-        this.isPowered = false;
         this.ticks = 0;
     }
 
     public void tick() {
-        if (this.ticks != 0 && this.ticks % 120 == 0) { // 60 ticks equals 3 seconds
+        if (this.ticks != 0 && this.ticks % 60 == 0) { // 60 ticks equals 3 seconds
             this.move();
         }
         this.ticks++;
@@ -51,61 +45,52 @@ public class DrillHeadBlockEntity extends InventoryBlockEntity {
 
     private void move() {
         if (!this.level.isClientSide) {
-            if (true) { // isPowered()
+            // System.out.println("DRILL BE POWERED : " + this.getBlockState().getValue(POWERED));
+            if (this.getBlockState().getValue(POWERED)) {
                 this.drillFrontBlocks();
-                this.moveToFront();
+                if (this.getBlockState().getValue(POWERED)) {
+                    //this.moveToFront();
+                }
             }
         }
     }
 
     private void moveToFront() {
-        this.getLevel().setBlock(this.getBlockPos(), Blocks.AIR.defaultBlockState(), 35);
-        this.getLevel().levelEvent(null, 2001, this.getBlockPos(), Block.getId(this.getBlockState()));
-
         BlockPos nextPos = this.getBlockPos().relative(this.getBlockState().getValue(FACING).getOpposite());
-        BlockPos upperNextPos = nextPos.above();
-                /*
-                switch (this.getBlockState().getValue(FACING)) {
-                    case NORTH:
-                        nextPos = nextPos.south();
-                        upperNextPos = upperNextPos.south();
-                        break;
-                    case EAST:
-                        nextPos = nextPos.west();
-                        upperNextPos = upperNextPos.west();
-                        break;
-                    case WEST:
-                        nextPos = nextPos.east();
-                        upperNextPos = upperNextPos.east();
-                        break;
-                    case SOUTH:
-                        nextPos = nextPos.north();
-                        upperNextPos = upperNextPos.north();
-                        break;
-                }
-                 */
-
-        this.getLevel().playSound(null, this.getBlockPos(), SoundEvents.PISTON_EXTEND, SoundSource.BLOCKS, 1.0F, 1.0F);
 
         this.getLevel().setBlock(nextPos, this.getBlockState(), 35);
-        this.getLevel().levelEvent(null, 2001, nextPos, Block.getId(this.getBlockState()));
 
-        // this.getLevel().setBlock(upperNextPos, this.getBlockState().setValue(HALF, DoubleBlockHalf.UPPER), 35);
-        // this.getLevel().levelEvent(null, 2001, upperNextPos, Block.getId(this.getBlockState()));
+        this.getLevel().destroyBlock(this.getBlockPos(), false);
+
+        this.getLevel().playSound(null, this.getBlockPos(), SoundEvents.PISTON_EXTEND, SoundSource.BLOCKS, 1.0F, 1.0F);
     }
 
     private void drillFrontBlocks() {
-        System.out.println("DRILL HEAD POS: " + this.getBlockPos());
         for (int i = -1; i < 2; i++) {
             for (int j = -1; j < 2; j++) {
                 // Drill a 3x3 zone of blocks
-                BlockPos destroyPos = this.getBlockPos().offset(-1, i, j);
-                // BlockPos destroyPos = new BlockPos(this.getBlockPos().getX() + j, this.getBlockPos().getY() + i, this.getBlockPos().getZ() + 1);
-                System.out.println("DESTROY POS: " + destroyPos);
+                System.out.println("DestroyPos : " + this.getBlockPos());
+                BlockPos destroyPos = this.getBlockPos();
+                switch (this.getBlockState().getValue(FACING).getOpposite()) {
+                    case NORTH:
+                        destroyPos = this.getBlockPos().offset(j, i, -1);
+                        break;
+                    case EAST:
+                        destroyPos = this.getBlockPos().offset(1, i, j);
+                        break;
+                    case WEST:
+                        destroyPos = this.getBlockPos().offset(-1, i, j);
+                        break;
+                    case SOUTH:
+                        destroyPos = this.getBlockPos().offset(j, i, 1);
+                        break;
+                }
 
-                this.getLevel().destroyBlock(destroyPos, true);
-
-                //this.getLevel().setBlock(destroyPos, Blocks.AIR.defaultBlockState(), 35);
+                if (this.getLevel().getBlockState(destroyPos).getBlock() == Blocks.BEDROCK) {
+                    this.getBlockState().setValue(POWERED, false);
+                } else {
+                    this.getLevel().destroyBlock(destroyPos, true);
+                }
 
                 /*
                 this.getLevel().setBlock(this.getBlockPos(), Blocks.AIR.defaultBlockState(), 35);
@@ -117,13 +102,5 @@ public class DrillHeadBlockEntity extends InventoryBlockEntity {
                  */
             }
         }
-    }
-
-    public boolean isPowered() {
-        return this.isPowered;
-    }
-
-    public void setPowered(boolean isPowered) {
-        this.isPowered = isPowered;
     }
 }
